requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
gradio>=4.0.0

